fix: 🐛 change height prop to work in media style props
Currently when writing height in a media size prop (sm,md,lg) it will
not work. The css will be included but will have a comma before it and
therefore not be read. This is due to media queried css being run
toString on in join and therefore adding commas after each item in the
css.